+++ test/jdk/jdk/incubator/vector/LongMaxVectorLoadStoreTests.java
@@ -803,7 +803,7 @@ public class LongMaxVectorLoadStoreTests extends AbstractVectorLoadStoreTest {
    @Test(dataProvider = "shuffleIntByteProviderForIOOBE")
    static void shuffleLoadMemorySegmentIOOBE(IntFunction<int[]> fa, IntFunction<Integer> fi) {
        MemorySegment a = toShuffleSegment(SPECIES, fa.apply(SPECIES.length()), i -> Arena.ofAuto().allocate(i, Integer.SIZE));
-       MemorySegment r = Arena.ofAuto().allocate(a.byteSize(), Integer.SIZE);
+       MemorySegment r = Arena.ofAuto().allocate(a.byteSize(), Long.SIZE);
 
        int l = (int) a.byteSize();
        int s = SPECIES.length() * 4;
@@ -832,7 +832,7 @@ public class LongMaxVectorLoadStoreTests extends AbstractVectorLoadStoreTest {
    @Test(dataProvider = "shuffleIntByteProviderForIOOBE")
    static void shuffleStoreMemorySegmentIOOBE(IntFunction<int[]> fa, IntFunction<Integer> fi) {
        MemorySegment a = toShuffleSegment(SPECIES, fa.apply(SPECIES.length()), i -> Arena.ofAuto().allocate(i, Integer.SIZE));
-       MemorySegment r = Arena.ofAuto().allocate(a.byteSize(), Integer.SIZE);
+       MemorySegment r = Arena.ofAuto().allocate(a.byteSize(), Long.SIZE);
 
        int l = (int) a.byteSize();
        int s = SPECIES.length() * 4;
